move DeviceTracker and JUnitTracker to local package
