+++ core/src/main/kotlin/com/malinskiy/marathon/analytics/local/DeviceTracker.kt
@@ -1,5 +1,6 @@
-package com.malinskiy.marathon.analytics
+package com.malinskiy.marathon.analytics.local
 
+import com.malinskiy.marathon.analytics.NoOpTracker
 import com.malinskiy.marathon.device.Device
 import com.malinskiy.marathon.device.DevicePoolId
 import com.malinskiy.marathon.report.internal.DeviceInfoReporter
